### 2019-11-05 #1860 @tomalrussell
Hi all - thanks for the speedy first steps.

Mentioning @edwardoughton to pull them into this issue for reference.

Out of curiosity -
