### 2019-11-05 #1860 @melissawm
@aaronyd, would you be willing to review this software paper for JOSS? Thank you!
