### 2019-11-21 #1860 @tomalrussell
Just looking through the JOSS reviewers list - perhaps @potterzot or @nmstreethran would be interested and available?
