### 2019-11-21 #1860 @nmstreethran
Hello @tomalrussell, I would be happy to review your paper.
