### 2019-11-21 #1860 @potterzot
@tomalrussell I would also be happy to review your paper, thanks!
