### 2019-11-22 #1911 @whedon
[REVIEW]: cdcam: Cambridge Digital Communications Assessment Model
**Submitting author:** @tomalrussell (<a href="http://orcid.org/0000-0002-0081-400X">Tom Russell</a>)
**Repository:** <a href="https://github.com/nismod/cdcam" target ="_blank">https://github.com/nismod/cdcam</a>
**Version:** v1.0.2
