### 2019-11-25 nismod/cdcam#14 @potterzot
Issues from JOSS review
Hi @tomalrussell - I'm filing this issue as part of my review for your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1911). The reviewer guidelines request that issues be filed at the package repository and then referenced from the review thread, so specific issues are detailed here and I will summarize in the submission thread. Below is my review response.

Overall the `cdcam` library looks nicely put together. My suggestions are mainly to do with documentation, and particularly with providing a working real-world example (even if substantially simplified) of analysis using the library, which I think would help new users understand how the library can be used in their own work.

Thank you for the cleanliness of the library code, it was a pleasure to review.

### Code and Tests

No issues with the code and test suite.

### Installation

It may be useful to include non-conda instructions for linux users who do not use anaconda. I used something like the following:

```
python -m venv cdcam
cdcam/bin/python cdcam/bin/pip install -r requirements.txt
cdcam/bin/python cdcam/bin/pip install -r requirements-dev.txt
cdcam/bin/python cdcam/bin/pip install cdcam
cdcam/bin/python cdcam/bin/pip install pytest pytest-cov
pytest --cov-report=term --cov=cdcam tests/
```

### Example

[Review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) state

> The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

The provided example in `scripts/run.py` generates a series of projections, but it would be helpful to have context in the `Getting Started` documentation about how these projections should be interpreted and what they contain. Currently `run.py` prints the results to the terminal for several categories. A discussion of what the categories are and what the outputs (i.e. "Service" and "Demand") mean would be helpful.

Alternatively, a more complete use case that demonstrates a question to be answered with the data at hand, the development of the model that is currently encapsulated in `scripts/run.py`, and a discussion of what the output of the model implies for the question would be a great help for new users. From the cited articles in this repository's paper I suspect there already exists an excellent example that could be stripped down and would work well.

This suggestion is different from what is currently in "Getting Started", which walks the user through the different parameters that need to be chosen. This is helpful, but I think would be more powerful and improve understanding if it was within the context of a real-world applicable use case that answers a question.

### Community Guidelines

Part of [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) indicate the need for

> clear guidelines for third-parties wishing to:
>   -    Contribute to the software
>   -    Report issues or problems with the software
>   -    Seek support

At the moment there is no information for these in the README or software documentation.

### Minor suggestions
- Some publicly facing methods lack examples in their documentation.
