### 2019-11-25 #1911 @potterzot
I filed a detailed review as an [issue on the cdcam repository](https://github.com/nismod/cdcam/issues/14) and have updated my checklist. Overall the library is very nicely put together. I suggest a couple of minor revisions having to do with providing non-conda installation instructions, a clear real-world example and providing information on contributing, seeking additional help, and filing issues that could be added to the README. Code executes and tests all pass.
