### 2019-12-02 nismod/cdcam#15 @nmstreethran
JOSS paper review
Hello @tomalrussell, I have finished reviewing your software paper submission to JOSS according to the checklist in https://github.com/openjournals/joss-reviews/issues/1911. Here are my comments.

### General checks

#### License

> Copyright (c) 2019 Edward Oughton

Should you be included in here as well, @tomalrussell?

### Functionality

I installed the software and performed the tests according to the instructions provided in `README.md` and had no issues. I also ran the example project using the data from Zenodo and there were no problems.

#### Code

There were a few typos in the docstrings in `model.py`.

- _urban_
https://github.com/nismod/cdcam/blob/e5ee9b1b1380ab64f038bc35949765611bf7e48a/src/cdcam/model.py#L81

- _District_
https://github.com/nismod/cdcam/blob/e5ee9b1b1380ab64f038bc35949765611bf7e48a/src/cdcam/model.py#L286

- _population_
https://github.com/nismod/cdcam/blob/e5ee9b1b1380ab64f038bc35949765611bf7e48a/src/cdcam/model.py#L600

The repository URL provided in `setup.py` seems outdated. It still redirects to the current repository though.
https://github.com/nismod/cdcam/blob/e5ee9b1b1380ab64f038bc35949765611bf7e48a/setup.py#L30

The code looks good otherwise!

### Documentation

The documentation does not clearly state the need for this software and the target audience. This should be similar to the statement of need in the paper. An example of how to use the software to solve real-world problems should also be provided here.

The installation instructions provided in the documentation is not as detailed as the one provided in `README.md`. Information about installing the package using conda could be included in the documentation. The review checklist asks for a clearly-stated list of dependencies, handled ideally with an automated package management solution.

The documentation does not contain community guidelines for third parties wishing to contribute to the software, report issues or problems with the software, and seek support. A [`CONTRIBUTING`](https://github.com/earthlab/earthpy/blob/master/CONTRIBUTING.rst) file should be included in the repository and documentation, and a link to the guidelines should be provided in `README.md`.

### Software paper

The last sentence of the 'Applications' section states the following, but no citations have been provided.
> Increasingly, the modelling capability reported here is being applied internationally, including to provide the analytics for the World Bank's Flagship 5G report.

The DOI fields in `paper.bib` should contain just the identifier and not the full URL, i.e., `doi = {10.1016/j.techfore.2018.03.016}`.

The paper is well written and can be understood easily by a non-specialist audience.
