### 2019-12-02 nismod/cdcam#15 @edwardoughton
- Added TR to licence
- Corrected typo - urban
- Corrected typo - district
- Corrected typo - population
- Corrected setup.py url
- Removed the reference to the WB 5G Flagship as it won't be released until 2020, hence there is no current DOI.
- Corrected the DOI fields in paper.bib

Still TODO:
- Documentation points 1 - 3
