### 2019-12-02 #1911 @nmstreethran
My apologies for accidentally closing the issue. I've completed my review and left my comments in https://github.com/nismod/cdcam/issues/15.
