### 2019-12-02 nismod/cdcam#15 @edwardoughton
- Completed documentation points 1 - 3
