### 2019-12-02 nismod/cdcam#14 @edwardoughton
- Added installation instructions
- Added community guidelines

Still TODO:
- Add better reprex examples
