### 2019-12-09 nismod/cdcam#14 @edwardoughton
Add scripts/reprex.py to provide self contained reproducible example which does not require the Zenodo data to be downloaded. This example reflects the data discussed in the documentation, so it provides additional information for the user.

Also added info on how to interpret the results in the documentation (as well as adding units).
