### 2019-12-15 nismod/cdcam#15 @nmstreethran
Thank you, @edwardoughton and @tomalrussell! I ran into some issues. I'm not sure if there's something wrong on my side as your automated tests seem to be working fine.

One of the tests failed due to assertion errors. The enhanced tests in [`test_mobile_interventions.py`](https://github.com/nismod/cdcam/blob/master/tests/test_mobile_interventions.py) are causing this. `macros_to_lte`, `macros_to_5g` and `small_cells` return zero instead. I'm also getting the `Could not create object for <postcode>` exception for this test and the new reproducible example.

I've included the outputs below.

<summary>pytest output</summary>

```
(cdcam) nms@nms:~/Documents/cdcam-master$ pytest --cov-report=term --cov=cdcam tests/
================================================== test session starts ==================================================
rootdir: /home/nms/Documents/cdcam-master
plugins: cov-2.8.1
collected 10 items

tests/test_mobile_interventions.py F                                                                              [ 10%]
tests/test_mobile_model.py .........                                                                              [100%]

======================================================= FAILURES ========================================================
_______________________________________________ test_decide_interventions _______________________________________________

non_4g_system = <cdcam.model.NetworkManager object at 0x7ffb0e054b10>
basic_system = <cdcam.model.NetworkManager object at 0x7ffb0e054c90>
mixed_system = <cdcam.model.NetworkManager object at 0x7ffb0e054d10>
empty_system = <cdcam.model.NetworkManager object at 0x7ffb0e054fd0>
high_demand_system = <cdcam.model.NetworkManager object at 0x7ffb0dfed150>
setup_simulation_parameters = {'annual_budget': 500000000, 'busy_hour_traffic_percentage': 15, 'channel_bandwidth_1800': '10', 'channel_bandwidth_2600': '10', ...}
setup_simulation_parameters2 = {'annual_budget': 1000000.0, 'busy_hour_traffic_percentage': 20, 'channel_bandwidth_1800': '10', 'channel_bandwidth_2600': '10', ...}

    def test_decide_interventions(non_4g_system, basic_system,
        mixed_system, empty_system, high_demand_system, setup_simulation_parameters,
        setup_simulation_parameters2):

        actual_result = decide_interventions(
            'minimal', 250000, 0,
            mixed_system, 2020, setup_simulation_parameters
        )

        assert actual_result == ([], 250000, [])

        actual_result = decide_interventions(
            'upgrade_to_lte', 142446, 2,
            non_4g_system, 2020, setup_simulation_parameters
        )

        assert len(actual_result[0]) == 2
        assert actual_result[1] == 0

        actual_result = decide_interventions(
            'upgrade_to_lte', 142446, 2,
            mixed_system, 2020, setup_simulation_parameters
        )

        assert actual_result == ([], 142446, [])

        # #50917 * 4 = 203668
        actual_result = decide_interventions(
            'macrocell', 109000, 1000,
            mixed_system, 2020, setup_simulation_parameters
        )

        assert len(actual_result[0]) == 4
        assert actual_result[1] == 0

        actual_result = decide_interventions(
            'macrocell', 203668, 0,
            mixed_system, 2020, setup_simulation_parameters
        )

        assert len(actual_result[0]) == 0
        assert actual_result[1] == 203668

        # #50917 * 2 = 101,834
        # #40220 * 3 = £120,660
        actual_result = decide_interventions(
            'small-cell-and-spectrum', 109000 , 1000,
            mixed_system, 2020, setup_simulation_parameters
        )

        assert len(actual_result[0]) == 4
        assert actual_result[1] == 0

        #test empty_system
        actual_result = decide_interventions(
            'small-cell-and-spectrum', 1e7 , 0,
            empty_system, 2020, setup_simulation_parameters2
        )

        macros_to_lte = len([a for a in actual_result[0] if  a['type'] == 'macrocell_site' \
                        and a['technology'] == 'LTE'])

>       assert macros_to_lte == 4
E       assert 0 == 4

tests/test_mobile_interventions.py:134: AssertionError
------------------------------------------------- Captured stdout setup -------------------------------------------------
could not create object for CB11
{'id': 'CB11', 'lad_id': 1, 'population': 100000, 'area_km2': 2, 'user_throughput': 2}
could not create object for CB12
{'id': 'CB12', 'lad_id': 1, 'population': 100000, 'area_km2': 2, 'user_throughput': 2}
could not create object for CB11
{'id': 'CB11', 'lad_id': 1, 'population': 100000, 'area_km2': 2, 'user_throughput': 2}
could not create object for CB12
{'id': 'CB12', 'lad_id': 1, 'population': 100000, 'area_km2': 2, 'user_throughput': 2}
------------------------------------------------- Captured stdout call --------------------------------------------------
Service 0
Demand 0
Service 2
Demand 0
Service 0
Demand 0
Service 4
Demand 0
Service 0
Demand 0
Service 4
Demand 0
Service 0
Demand 0

----------- coverage: platform linux, python 3.7.5-final-0 -----------
Name                                                                                Stmts   Miss  Cover
-------------------------------------------------------------------------------------------------------
/home/nms/anaconda3/envs/cdcam/lib/python3.7/site-packages/cdcam/__init__.py            6      2    67%
/home/nms/anaconda3/envs/cdcam/lib/python3.7/site-packages/cdcam/interventions.py     168     48    71%
/home/nms/anaconda3/envs/cdcam/lib/python3.7/site-packages/cdcam/model.py             169      7    96%
-------------------------------------------------------------------------------------------------------
TOTAL                                                                                 343     57    83%

============================================== 1 failed, 9 passed in 0.19s ==============================================
```

<summary>example.py output</summary>

```
(cdcam) nms@nms:~/Documents/cdcam-master$ python scripts/example.py

Running: baseline baseline small-cell-and-spectrum

----------------------------------------------------
- 2020 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

could not create object for CB11
{'id': 'CB11', 'lad_id': 'E07000008', 'population': 100000, 'area_km2': 2, 'user_throughput': 1}
could not create object for CB12
{'id': 'CB12', 'lad_id': 'E07000008', 'population': 200000, 'area_km2': 2, 'user_throughput': 1}
Service 0
Demand 0
could not create object for CB11
{'id': 'CB11', 'lad_id': 'E07000008', 'population': 100000, 'area_km2': 2, 'user_throughput': 1}
could not create object for CB12
{'id': 'CB12', 'lad_id': 'E07000008', 'population': 200000, 'area_km2': 2, 'user_throughput': 1}

Built 0 new assets in 2020:
-- 0 LTE Macro Cells
-- 0 5G Macro Cells
-- 0 5G Small Cells

Financials
----------
£££ - Spent £0.0 million
£££ - Budget remaining £1.0 million

Cambridge:
-- Demand (Mbps/km²): 0,
-- Capacity (Mbps/km²): 0

----------------------------------------------------
- 2021 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

Service 0
Demand 0
could not create object for CB11
{'id': 'CB11', 'lad_id': 'E07000008', 'population': 110000, 'area_km2': 2, 'user_throughput': 2}
could not create object for CB12
{'id': 'CB12', 'lad_id': 'E07000008', 'population': 220000, 'area_km2': 2, 'user_throughput': 2}

Built 0 new assets in 2021:
-- 0 LTE Macro Cells
-- 0 5G Macro Cells
-- 0 5G Small Cells

Financials
----------
£££ - Spent £0.0 million
£££ - Budget remaining £1.0 million

Cambridge:
-- Demand (Mbps/km²): 0,
-- Capacity (Mbps/km²): 0

----------------------------------------------------
- 2022 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

Service 0
Demand 0
could not create object for CB11
{'id': 'CB11', 'lad_id': 'E07000008', 'population': 121000, 'area_km2': 2, 'user_throughput': 3}
could not create object for CB12
{'id': 'CB12', 'lad_id': 'E07000008', 'population': 232000, 'area_km2': 2, 'user_throughput': 3}

Built 0 new assets in 2022:
-- 0 LTE Macro Cells
-- 0 5G Macro Cells
-- 0 5G Small Cells

Financials
----------
£££ - Spent £0.0 million
£££ - Budget remaining £1.0 million

Cambridge:
-- Demand (Mbps/km²): 0,
-- Capacity (Mbps/km²): 0
```

---

**UPDATE:** I don't get these errors when I clone the repository and run `python setup.py develop`. So this was probably just caused by the out-of-date package on PyPI.

<summary>pytest output</summary>

```
(cdcam-dev) nms@nms:~/Documents/cdcam-master$ pytest --cov-report=term --cov=cdcam tests/
======================================================= test session starts ========================================================
rootdir: /home/nms/Documents/cdcam-master
plugins: cov-2.8.1
collected 10 items

tests/test_mobile_interventions.py .                                                                                         [ 10%]
tests/test_mobile_model.py .........                                                                                         [100%]

----------- coverage: platform linux, python 3.7.5-final-0 -----------
Name                         Stmts   Miss  Cover
------------------------------------------------
src/cdcam/__init__.py            6      2    67%
src/cdcam/interventions.py     167     10    94%
src/cdcam/model.py             172     11    94%
------------------------------------------------
TOTAL                          345     23    93%
```

<summary>example.py output</summary>

```
(cdcam-dev) nms@nms:~/Documents/cdcam-master$ python scripts/example.py

Running: baseline baseline small-cell-and-spectrum

----------------------------------------------------
- 2020 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

Built 6 assets to meet service obligation, 10 to meet demand

Built 16 new assets in 2020:
-- 8 LTE Macro Cells
-- 0 5G Macro Cells
-- 8 5G Small Cells

Financials
----------
£££ - Spent £0.7 million
£££ - Budget remaining £0.3 million

Cambridge:
-- Demand (Mbps/km²): 273,
-- Capacity (Mbps/km²): 400

CB11:
-- Demand (Mbps/km²): 182,
-- Capacity (Mbps/km²): 350
CB12:
-- Demand (Mbps/km²): 364,
-- Capacity (Mbps/km²): 450

----------------------------------------------------
- 2021 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

Built 0 assets to meet service obligation, 0 to meet demand

Built 0 new assets in 2021:
-- 0 LTE Macro Cells
-- 0 5G Macro Cells
-- 0 5G Small Cells

Financials
----------
£££ - Spent £0.0 million
£££ - Budget remaining £1.0 million

Cambridge:
-- Demand (Mbps/km²): 601,
-- Capacity (Mbps/km²): 400

CB11:
-- Demand (Mbps/km²): 400,
-- Capacity (Mbps/km²): 350
CB12:
-- Demand (Mbps/km²): 801,
-- Capacity (Mbps/km²): 450

----------------------------------------------------
- 2022 ( baseline baseline small-cell-and-spectrum )
----------------------------------------------------

Built 0 assets to meet service obligation, 55 to meet demand

Built 55 new assets in 2022:
-- 0 LTE Macro Cells
-- 3 5G Macro Cells
-- 52 5G Small Cells

Financials
----------
£££ - Spent £1.0 million
£££ - Budget remaining £-0.0 million

Cambridge:
-- Demand (Mbps/km²): 964,
-- Capacity (Mbps/km²): 475

CB11:
-- Demand (Mbps/km²): 661,
-- Capacity (Mbps/km²): 350
CB12:
-- Demand (Mbps/km²): 1267,
-- Capacity (Mbps/km²): 600
```

---

### Documentation

[index.rst](https://github.com/nismod/cdcam/blob/master/docs/index.rst#applications) includes some citations using the Pandoc syntax. These should be removed or replaced with hyperlinks pointing to the sources.

Also there's one last typo in line 41 of [getting-started.rst](https://github.com/nismod/cdcam/blob/master/docs/getting-started.rst).

> A major motivation for this scenario analysis is to examine how changes in population and urban development might impact on infrastructure service demands, and what might need to be done to meet those needs or **expecations**.
