### 2019-12-16 nismod/cdcam#15 @tomalrussell
Thanks @nmstreethran !

- linked citations in index.rst: https://cdcam.readthedocs.io/en/latest/
- fixed that typo: https://github.com/nismod/cdcam/commit/031e32afa5d89c02fb7b596a99c0960e3e63be5d
- updated PyPI package to 1.0.1 https://pypi.org/project/cdcam/ - so `pip install --upgrade cdcam` should work as well as testing from repository checkout.
