### 2019-12-16 #1911 @nmstreethran
Hi @melissawm, I've updated my review checklist and am happy with the revisions by @tomalrussell. I think this submission is ready for publication. :)
