### 2020-01-27 #1911 @melissawm
@openjournals/joss-eics looks like this is ready :)
