### 2020-01-29 #1911 @Kevin-Mattheus-Moerman
@tomalrussell I'm the AEiC handling the processing for this submission. Below are some remaining minor points to work on. Let me know when you've processed these.

- [ ] Please ensure the [archive](https://zenodo.org/record/3583132#.XjFGZeGnxKg) meta data matches that for the paper i.e. the paper title and the author name(s) should be the same
- [ ]  Can you confirm that the version tag v1.0.2 is the most accurate for this paper/review/archive?
- [ ] I've proofread the paper and it reads well. This paper is about to be accepted for publication in JOSS. Please double check the paper yourself at this point as well.
