### 2020-01-29 #1911 @tomalrussell
Thanks @Kevin-Mattheus-Moerman - I think this is all okay:
- [x] archive and paper are both Oughton and Russell
- [x] v1.0.2 is the most accurate version tag
- [x] I've read through the paper again
