### 2020-01-29 #1911 @tomalrussell
@Kevin-Mattheus-Moerman I've now edited the archive metadata, title is now "cdcam: Cambridge Digital Communications Assessment Model".
